# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "POSCOSEGRAN API",
    "description": "Sistema experto para condiciones de almacenamiento poscosecha de maíz chulpi. Evalúa condiciones de almacenamiento; no certifica inocuidad, ausencia de micotoxinas ni aptitud para consumo.",
    "version": "0.7.0"
  },
  "paths": {
...
    "/api/v1/unidades/{id_unidad}/borrador": {
      "get": {
        "tags": [
          "evaluacion"
        ],
        "summary": "Obtener Borrador",
        "operationId": "obtener_borrador_api_v1_unidades__id_unidad__borrador_get",
        "parameters": [
          {
            "name": "id_unidad",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id Unidad"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Borrador"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "evaluacion"
        ],
        "summary": "Guardar Borrador",
        "description": "Un borrador no autoriza nada ni invalida el resultado vigente.\n\nComunicar un hallazgo adverso real exige enviarlo como evaluación o control.",
        "operationId": "guardar_borrador_api_v1_unidades__id_unidad__borrador_put",
        "parameters": [
          {
            "name": "id_unidad",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id Unidad"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          },
          {
            "name": "If-None-Match",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-None-Match"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluacionEntrada-Input"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Borrador"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/unidades/{id_unidad}/evaluaciones": {
      "post": {
        "tags": [
          "evaluacion"
        ],
        "summary": "Evaluar",
        "description": "Emite una evaluación. SIN_CONCLUSION_AUTOMATICA es una respuesta exitosa.",
        "operationId": "evaluar_api_v1_unidades__id_unidad__evaluaciones_post",
        "parameters": [
          {
            "name": "id_unidad",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id Unidad"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluacionEntrada-Input"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultadoEvaluacion"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evaluaciones/{id_evaluacion}": {
      "get": {
        "tags": [
          "evaluacion"
        ],
        "summary": "Obtener Evaluacion",
        "operationId": "obtener_evaluacion_api_v1_evaluaciones__id_evaluacion__get",
        "parameters": [
          {
            "name": "id_evaluacion",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id Evaluacion"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultadoEvaluacion"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/unidades/{id_unidad}/historial": {
      "get": {
        "tags": [
          "evaluacion"
        ],
        "summary": "Historial",
        "operationId": "historial_api_v1_unidades__id_unidad__historial_get",
        "parameters": [
          {
            "name": "id_unidad",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id Unidad"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limite"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pagina_ResultadoEvaluacion_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/unidades/{id_unidad}/vigencia": {
      "get": {
        "tags": [
          "evaluacion"
        ],
        "summary": "Consultar Vigencia",
        "operationId": "consultar_vigencia_api_v1_unidades__id_unidad__vigencia_get",
        "parameters": [
          {
            "name": "id_unidad",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id Unidad"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vigencia"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Borrador": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "id_unidad": {
            "type": "string",
            "format": "uuid",
            "title": "Id Unidad"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "contenido": {
            "$ref": "#/components/schemas/EvaluacionEntrada-Output"
          },
          "actualizado_en": {
            "type": "string",
            "format": "date-time",
            "title": "Actualizado En"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "id_unidad",
          "revision",
          "contenido",
          "actualizado_en"
        ],
        "title": "Borrador"
      },
...
      "EvaluacionEntrada-Input": {
        "properties": {
          "revision_unidad": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Revision Unidad"
          },
          "revision_almacen": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Revision Almacen"
          },
          "fase": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "INGRESO",
                  "SEGUIMIENTO"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Fase"
          },
          "observaciones": {
            "items": {
              "$ref": "#/components/schemas/ObservacionEntrada"
            },
            "type": "array",
            "title": "Observaciones"
          },
          "historial": {
            "$ref": "#/components/schemas/HistorialEntrada"
          },
          "dias_previstos_restantes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dias Previstos Restantes"
          },
          "fecha_salida_prevista": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Salida Prevista"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "revision_unidad",
          "revision_almacen",
          "fase",
          "observaciones",
          "historial",
          "dias_previstos_restantes",
          "fecha_salida_prevista"
        ],
        "title": "EvaluacionEntrada"
      },
      "EvaluacionEntrada-Output": {
        "properties": {
          "revision_unidad": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Revision Unidad"
          },
          "revision_almacen": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Revision Almacen"
          },
          "fase": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "INGRESO",
                  "SEGUIMIENTO"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Fase"
          },
          "observaciones": {
            "items": {
              "$ref": "#/components/schemas/ObservacionEntrada"
            },
            "type": "array",
            "title": "Observaciones"
          },
          "historial": {
            "$ref": "#/components/schemas/HistorialEntrada"
          },
          "dias_previstos_restantes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dias Previstos Restantes"
          },
          "fecha_salida_prevista": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Salida Prevista"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "revision_unidad",
          "revision_almacen",
          "fase",
          "observaciones",
          "historial",
          "dias_previstos_restantes",
          "fecha_salida_prevista"
        ],
        "title": "EvaluacionEntrada"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HistorialEntrada": {
        "properties": {
          "fecha_inicio_historial": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Inicio Historial"
          },
          "vida_previa_documentada": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vida Previa Documentada"
          },
          "evidencia_vida_previa": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidencia Vida Previa"
          },
          "intervalos_historial": {
            "items": {
              "$ref": "#/components/schemas/IntervaloEntrada"
            },
            "type": "array",
            "title": "Intervalos Historial"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "fecha_inicio_historial",
          "vida_previa_documentada",
          "evidencia_vida_previa",
          "intervalos_historial"
        ],
        "title": "HistorialEntrada"
      },
...
      "Pagina_ResultadoEvaluacion_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ResultadoEvaluacion"
            },
            "type": "array",
            "title": "Items"
          },
          "siguiente_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Siguiente Cursor"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "siguiente_cursor"
        ],
        "title": "Pagina[ResultadoEvaluacion]"
      },
...
      "ResultadoEvaluacion": {
        "properties": {
          "evaluacion": {
            "$ref": "#/components/schemas/Evaluacion"
          },
          "vigencia": {
            "$ref": "#/components/schemas/Vigencia"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "evaluacion",
          "vigencia"
        ],
        "title": "ResultadoEvaluacion"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Vigencia": {
        "properties": {
          "id_unidad": {
            "type": "string",
            "format": "uuid",
            "title": "Id Unidad"
          },
          "id_evaluacion": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id Evaluacion"
          },
          "estado": {
            "type": "string",
            "enum": [
              "VIGENTE",
              "VENCIDA",
              "INVALIDADA",
              "NO_AUTORIZADO",
              "SIN_EVALUACION"
            ],
            "title": "Estado"
          },
          "consultada_en": {
            "type": "string",
            "format": "date-time",
            "title": "Consultada En"
          },
          "fecha_proximo_control": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Proximo Control"
          },
          "fecha_vencimiento_autorizacion": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Vencimiento Autorizacion"
          },
          "causas": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Causas"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id_unidad",
          "id_evaluacion",
          "estado",
          "consultada_en",
          "fecha_proximo_control",
          "fecha_vencimiento_autorizacion",
          "causas"
        ],
        "title": "Vigencia"
      }
    }
  }
}